requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pymongo==4.12.0
dnspython==2.1.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.5.0
sqlparse==0.5.3
certifi